+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add attachments capabilities
 - Updated github repo template assets
 - Fix upload error for users not assigned to a group
+- Add default publication workflow
 
 ## 1.1.0